feat: config eslint and husky
